upgrade redspot  (#35)
* name

* update

* update

* update

* update

* readme

+++ packages/redspot-core/src/builtin-tasks/clean.ts
@@ -1,17 +1,21 @@
-import fsExtra from "fs-extra";
-import { task } from "../internal/core/config/config-env";
-import { TASK_CLEAN } from "./task-names";
+import fsExtra from 'fs-extra';
+import { subtask, task } from '../internal/core/config/config-env';
+import { getCacheDir } from '../internal/util/global-dir';
+import { TASK_CLEAN, TASK_CLEAN_GLOBAL } from './task-names';
 
-export default function () {
-  task(
-    TASK_CLEAN,
-    "Clears the cache and deletes all artifacts",
-    async (_, { config }) => {
-      console.log(`Delete ${config.paths.cache}`);
-      await fsExtra.emptyDir(config.paths.cache);
-      console.log(`Delete ${config.paths.artifacts}`);
-      await fsExtra.remove(config.paths.artifacts);
-      console.log("Done!");
+subtask(TASK_CLEAN_GLOBAL, async () => {
+  const globalCacheDir = await getCacheDir();
+
+  await fsExtra.emptyDir(globalCacheDir);
+});
+
+task(TASK_CLEAN, 'Clears the cache and deletes all artifacts')
+  .addFlag('global', 'Clear the global cache')
+  .setAction(async ({ global }: { global: boolean }, { config, run }) => {
+    if (global) {
+      return run(TASK_CLEAN_GLOBAL);
     }
-  );
-}
+
+    await fsExtra.emptyDir(config.paths.cache);
+    await fsExtra.remove(config.paths.artifacts);
+  });
